+++ source/entity.cpp
@@ -64,7 +64,7 @@ bool Entity::Initialise(Renderer& renderer, const char* spritePath, int spriteSi
     } while (IsInsideWall(m_position));
 
     m_hitbox.setPosition(m_position);
-    m_hitbox.setDimensions(m_pSprite->GetWidth(), m_pSprite->GetHeight());
+    m_hitbox.setDimensions((float)m_pSprite->GetWidth(), (float)m_pSprite->GetHeight());
 
     return (m_pSprite != nullptr);
 }
