+++ source/player.cpp
@@ -24,7 +24,7 @@ bool Player::Initialise(Renderer& renderer)
 	Entity::Initialise(renderer, "sprites\\player.png", 300); // super();
 
 	m_pSprite->SetScale(m_fInitialScale);
-	m_hitbox.setDimensions(m_pSprite->GetWidth(), m_pSprite->GetHeight());
+	m_hitbox.setDimensions((float)m_pSprite->GetWidth(), (float)m_pSprite->GetHeight());
 
 	const int SCREEN_WIDTH = renderer.GetWidth();
 	const int SCREEN_HEIGHT = renderer.GetHeight();
